requirements.txt
colorama==0.4.4
Flask==2.1.2
pycryptodome==3.14.1
requests==2.25.1
tinydb==4.7.0
pyinstaller==5.1